adding route from signup to login

+++ src/utils/authStorage.ts
@@ -1,5 +1,5 @@
 export interface User {
-  id:number
+  id? :number
   fullname : string;
   username: string;
   password: string;
@@ -10,12 +10,19 @@ export const getUser= ():User[]=>{
    return JSON.parse(localStorage.getItem("users")||"[]")
 }
 
-export const addUser = (user:User):void => {
+export const addUser = (user: User): boolean => {
   const users = getUser();
-  users.push(user);
-  localStorage.setItem("users",JSON.stringify(users));
-}
 
+  const alreadyExists = users.some((u) => u.username === user.username);
+
+  if (alreadyExists) {
+    return false;
+  }
+
+  users.push(user);
+  localStorage.setItem("users", JSON.stringify(users));
+  return true;
+};
 
 export const findUser = (username: string, password: string): User | undefined => {
   const users = getUser();
